requirements.txt
amqp==5.1.1
asgiref==3.7.2
astroid==2.15.6
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.5
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.2
defusedxml==0.7.1
dill==0.3.6
dj-database-url==2.0.0
Django==4.2.3
django-debug-toolbar==4.1.0
django-filter==23.2
django-redis==5.3.0
django-rich-logging==0.3.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-nested-routers==0.93.4
drf-spectacular==0.26.3
exceptiongroup==1.1.2
flower==2.0.0
gunicorn==21.0.1
humanize==4.7.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
install==1.3.5
isort==5.12.0
jsonschema==4.18.4
jsonschema-specifications==2023.6.1
kombu==5.3.1
lazy-object-proxy==1.9.0
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
mysqlclient==2.2.0
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.0
platformdirs==3.8.1
pluggy==1.2.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg==3.1.9
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.15.1
PyJWT==2.7.0
pylint==2.17.4
pyrsistent==0.19.3
pytest==7.4.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
referencing==0.29.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.4.2
rpds-py==0.8.11
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.11.8
tornado==6.3.2
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.5.0
wrapt==1.15.0
